Move disk-independent parts of openssl_privatekey to module_utils and doc_fragments.
